+++ src/sys/io/define_no_std/cursor.rs
@@ -1,7 +1,48 @@
 // devela::sys::io::reimplement_no_std::cursor
+//
+//! Defines [`IoCursor`], [`IoSeek`], [`IoSeekFrom`].
+//
+// TOC
+// - trait IoSeek
+// - enum IoSeekFrom
+// - struct IoCursor
 
-use crate::{IoBufRead, IoError, IoErrorKind, IoRead, IoResult, IoSeek, IoSeekFrom, IoWrite};
+use crate::{IoBufRead, IoError, IoErrorKind, IoRead, IoResult, IoWrite};
 use ::core::cmp;
+
+/// The `IoSeek` trait provides a cursor which can be moved within a stream of
+/// bytes.
+///
+/// This struct is generally created by calling [`bytes`][IoRead::bytes] on a reader.
+///
+/// See <https://doc.rust-lang.org/std/io/trait.Seek.html>.
+pub trait IoSeek {
+    /// Seek to an offset, in bytes, in a stream.
+    fn seek(&mut self, pos: IoSeekFrom) -> IoResult<u64>;
+}
+
+/// Enumeration of possible methods to seek within an I/O object.
+///
+/// It is used by the [`IoSeek`] trait.
+#[derive(Copy, PartialEq, Eq, Clone, Debug)]
+pub enum IoSeekFrom {
+    /// Sets the offset to the provided number of bytes.
+    Start(u64),
+
+    /// Sets the offset to the size of this object plus the specified number of
+    /// bytes.
+    ///
+    /// It is possible to seek beyond the end of an object, but it's an error to
+    /// seek before byte 0.
+    End(i64),
+
+    /// Sets the offset to the current position plus the specified number of
+    /// bytes.
+    ///
+    /// It is possible to seek beyond the end of an object, but it's an error to
+    /// seek before byte 0.
+    Current(i64),
+}
 
 /// Wraps an in-memory buffer and provides it with an [`IoSeek`] implementation.
 ///
@@ -14,59 +55,32 @@ pub struct IoCursor<T> {
 #[rustfmt::skip]
 impl<T> IoCursor<T> {
     /// Creates a new cursor wrapping the provided underlying in-memory buffer.
-    ///
-    /// See <https://doc.rust-lang.org/std/io/struct.Cursor.html#method.new>.
     pub fn new(inner: T) -> IoCursor<T> { IoCursor { pos: 0, inner } }
-
     /// Consumes this cursor, returning the underlying value.
-    ///
-    /// See <https://doc.rust-lang.org/std/io/struct.Cursor.html#method.into_inner>.
     pub fn into_inner(self) -> T { self.inner }
-
     /// Gets a reference to the underlying value in this cursor.
-    ///
-    /// See <https://doc.rust-lang.org/std/io/struct.Cursor.html#method.get_ref>.
     pub fn get_ref(&self) -> &T { &self.inner }
-
     /// Gets a mutable reference to the underlying value in this cursor.
-    ///
-    /// See <https://doc.rust-lang.org/std/io/struct.Cursor.html#method.get_mut>.
     pub fn get_mut(&mut self) -> &mut T { &mut self.inner }
-
     /// Returns the current position of this cursor.
-    ///
-    /// See <https://doc.rust-lang.org/std/io/struct.Cursor.html#method.position>.
     pub fn position(&self) -> u64 { self.pos }
-
     /// Sets the position of this cursor.
-    ///
-    /// See <https://doc.rust-lang.org/std/io/struct.Cursor.html#method.set_position>.
     pub fn set_position(&mut self, pos: u64) { self.pos = pos; }
 }
-
+#[rustfmt::skip]
 impl<T: AsRef<[u8]>> IoSeek for IoCursor<T> {
     fn seek(&mut self, style: IoSeekFrom) -> IoResult<u64> {
         let (base_pos, offset) = match style {
-            IoSeekFrom::Start(n) => {
-                self.pos = n;
-                return Ok(n);
-            }
+            IoSeekFrom::Start(n) => { self.pos = n; return Ok(n); }
             IoSeekFrom::End(n) => (self.inner.as_ref().len() as u64, n),
             IoSeekFrom::Current(n) => (self.pos, n),
         };
-        let new_pos = if offset >= 0 {
-            base_pos.checked_add(offset as u64)
-        } else {
-            base_pos.checked_sub((offset.wrapping_neg()) as u64)
-        };
+        let new_pos = if offset >= 0 { base_pos.checked_add(offset as u64) }
+            else { base_pos.checked_sub((offset.wrapping_neg()) as u64) };
         match new_pos {
-            Some(n) => {
-                self.pos = n;
-                Ok(self.pos)
-            }
+            Some(n) => { self.pos = n; Ok(self.pos) }
             None => Err(IoError::new(
-                IoErrorKind::InvalidInput,
-                "invalid seek to a negative or overflowing position",
+                IoErrorKind::InvalidInput, "invalid seek to a negative or overflowing position",
             )),
         }
     }